Finished divided components in GameOver.js and finished first structure
