+++ src/components/static/avatarList.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const avatars = [
+    {value: 'bardock',  text: 'Bardock'},
+    {value: 'c18',      text: 'C18'},
+    {value: 'gokussj3', text: 'Goku Ssj3'},
+    {value: 'vegeta',   text: 'Vegeta'} ];
+
+const avatarList = avatars.map( char => {
+    return <option value={char.value}>{ char.text }</option>
+});
+
+export default avatarList;
