+++ src/components/static/createBy.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const createBy = (
+    <div className="create-by">
+        <p>Game create by <span><i>Marek Ficht</i></span></p>
+        <p>All rights reserved &copy;</p>
+    </div>
+);
+
+export default createBy;
