+++ src/components/static/logo.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const logo = <h1 className="logo">Reflex game</h1>;
+
+export default logo;
