+++ src/components/static/randomChar.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const randomChar = ['X', 'Y', 'Z'];
+
+export default randomChar;
