+++ src/components/static/waitingForPlayers.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const waitingForPlayers = (
+    <div className="connection-info">
+        <p>Oczekiwanie na polaczanie</p>
+        <p>gracza</p>
+    </div>
+);
+
+export default waitingForPlayers;
