+++ src/styles/components/_Game.scss
@@ -145,8 +145,8 @@
 
 @keyframes prepare-count {
   from { transform: scale(0.75); }
-  20% { transform: scale(0.75); }
+  10% { transform: scale(0.75); }
   50% { transform: scale(1.2); }
-  85% { transform: scale(0.75); }
+  90% { transform: scale(0.75); }
   to { transform: scale(0.75); }
 }
